+++ sia/persist/sqlite/objects.go
@@ -15,18 +15,13 @@ func (s *Store) DeleteObject(accessKeyID, bucket, name string) error {
 		if err != nil {
 			return err
 		}
-		_, err = tx.Exec("DELETE FROM objects WHERE bucket_id = $1 AND name = $2 ", bid, name)
+		_, err = tx.Exec("DELETE FROM objects WHERE bucket_id = $1 AND name = $2", bid, name)
 		return err
 	})
 }
 
 // GetObject retrieves the object with the given bucket and name.
 func (s *Store) GetObject(accessKeyID *string, bucket, name string) (slabs.SealedObject, error) {
-	if accessKeyID == nil {
-		// public buckets not supported yet
-		return slabs.SealedObject{}, s3errs.ErrAccessDenied
-	}
-
 	var encoded []byte
 	err := s.transaction(func(tx *txn) error {
 		bid, err := bucketID(tx, bucket)
